refactor: resolve feedback points
