+++ src/courseware/course/sidebar/sidebars/course-outline/CourseOutlineTray.scss
