+++ src/courseware/course/sidebar/sidebars/course-outline/CourseOutlineTrigger.jsx
