+++ src/courseware/course/sidebar/sidebars/course-outline/components/CompletionIcon.jsx
