+++ src/courseware/course/sidebar/sidebars/course-outline/components/CompletionIcon.test.jsx
