+++ src/courseware/course/sidebar/sidebars/course-outline/components/SidebarSection.jsx
