+++ src/courseware/course/sidebar/sidebars/course-outline/components/SidebarSection.test.jsx
@@ -6,7 +6,7 @@ import { AppProvider } from '@edx/frontend-platform/react';
 
 import { initializeTestStore } from '@src/setupTest';
 import courseOutlineMessages from '@src/course-home/outline-tab/messages';
-import SidebarContext from '@src/courseware/course/sidebar/SidebarContext';
+import SidebarContext from '../../../SidebarContext';
 import SidebarSection from './SidebarSection';
 
 describe('<SidebarSection />', () => {
