+++ src/courseware/course/sidebar/sidebars/course-outline/components/SidebarSequence.jsx
