+++ src/courseware/course/sidebar/sidebars/course-outline/components/UnitIcon.test.jsx
