+++ src/courseware/course/sidebar/sidebars/course-outline/icons/DashedCircleIcon.jsx
