+++ src/courseware/course/sidebar/sidebars/course-outline/icons/index.js
