+++ src/index.scss
@@ -469,4 +469,4 @@
 @import "course-home/courseware-search/courseware-search.scss";
 @import "course-tabs/course-tabs-navigation.scss";
 @import "courseware/course/sidebar/common/SidebarBase.scss";
-@import "widgets/course-outline/CourseOutlineTray.scss";
+@import "courseware/course/sidebar/sidebars/course-outline/CourseOutlineTray.scss";
